pyproject.toml
[project]
name = "hdb-rag"
version = "0.1.0"
description = "RAG chatbot over Singapore HDB housing rules"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "langchain-chroma>=0.1.4",
    "langchain-postgres>=0.0.12",
    "chromadb>=0.5.0",
    "psycopg[binary]>=3.2.0",
    "streamlit>=1.38.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12.0",
    "pypdf>=4.0.0",
    "rank-bm25>=0.2.2",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "responses>=0.25.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hdb_rag"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
